Add websockets to testclient [skip ci]
