+++ lilya/testclient/base.py
@@ -37,14 +37,11 @@ from __future__ import annotations
 import contextlib
 import inspect
 import io
-import json
 import math
-import queue
 import sys
 import typing
 import warnings
 from concurrent.futures import Future
-from functools import cached_property
 from types import GeneratorType
 from urllib.parse import unquote, urljoin
 
@@ -54,8 +51,9 @@ from anyio.abc import ObjectReceiveStream, ObjectSendStream
 from anyio.streams.stapled import StapledObjectStream
 
 from lilya.compat import is_async_callable
+from lilya.testclient.exceptions import UpgradeException
+from lilya.testclient.websockets import WebSocketTestSession
 from lilya.types import ASGIApp, Message, Receive, Scope, Send
-from lilya.websockets import WebSocketDisconnect
 
 try:
     import httpx
@@ -105,167 +103,6 @@ class _WrapASGI2:
 class _AsyncBackend(typing.TypedDict):
     backend: str
     backend_options: dict[str, typing.Any]
-
-
-class _Upgrade(Exception):
-    def __init__(self, session: WebSocketTestSession) -> None:
-        self.session = session
-
-
-class WebSocketDenialResponse(
-    httpx.Response,
-    WebSocketDisconnect,
-):
-    """
-    A special case of `WebSocketDisconnect`, raised in the `TestClient` if the
-    `WebSocket` is closed before being accepted with a `send_denial_response()`.
-    """
-
-    ...
-
-
-class WebSocketTestSession:
-    def __init__(
-        self,
-        app: ASGI3App,
-        scope: Scope,
-        portal_factory: _PortalFactoryType,
-    ) -> None:
-        self.app = app
-        self.scope = scope
-        self.accepted_subprotocol = None
-        self.portal_factory = portal_factory
-        self._receive_queue: queue.Queue[Message] = queue.Queue()
-        self._send_queue: queue.Queue[Message | BaseException] = queue.Queue()
-        self.extra_headers = None
-
-    def __enter__(self) -> WebSocketTestSession:
-        self.exit_stack = contextlib.ExitStack()
-        self.portal = self.exit_stack.enter_context(self.portal_factory())
-
-        try:
-            _: Future[None] = self.portal.start_task_soon(self._run)
-            self.send({"type": "websocket.connect"})
-            message = self.receive()
-            self._raise_on_close(message)
-        except Exception:
-            self.exit_stack.close()
-            raise
-        self.accepted_subprotocol = message.get("subprotocol", None)
-        self.extra_headers = message.get("headers", None)
-        return self
-
-    @cached_property
-    def should_close(self) -> anyio.Event:
-        return anyio.Event()
-
-    async def _notify_close(self) -> None:
-        self.should_close.set()
-
-    def __exit__(self, *args: typing.Any) -> None:
-        try:
-            self.close(1000)
-        finally:
-            self.portal.start_task_soon(self._notify_close)
-            self.exit_stack.close()
-        while not self._send_queue.empty():
-            message = self._send_queue.get()
-            if isinstance(message, BaseException):
-                raise message
-
-    async def _run(self) -> None:
-        """
-        The sub-thread in which the websocket session runs.
-        """
-
-        async def run_app(tg: anyio.abc.TaskGroup) -> None:
-            try:
-                await self.app(self.scope, self._asgi_receive, self._asgi_send)
-            except anyio.get_cancelled_exc_class():
-                ...
-            except BaseException as exc:
-                self._send_queue.put(exc)
-                raise
-            finally:
-                tg.cancel_scope.cancel()
-
-        async with anyio.create_task_group() as tg:
-            tg.start_soon(run_app, tg)
-            await self.should_close.wait()
-            tg.cancel_scope.cancel()
-
-    async def _asgi_receive(self) -> Message:
-        while self._receive_queue.empty():
-            await anyio.sleep(0)
-        return self._receive_queue.get()
-
-    async def _asgi_send(self, message: Message) -> None:
-        self._send_queue.put(message)
-
-    def _raise_on_close(self, message: Message) -> None:
-        if message["type"] == "websocket.close":
-            raise WebSocketDisconnect(
-                code=message.get("code", 1000), reason=message.get("reason", "")
-            )
-        elif message["type"] == "websocket.http.response.start":
-            status_code: int = message["status"]
-            headers: list[tuple[bytes, bytes]] = message["headers"]
-            body: list[bytes] = []
-            while True:
-                message = self.receive()
-                assert message["type"] == "websocket.http.response.body"
-                body.append(message["body"])
-                if not message.get("more_body", False):
-                    break
-            raise WebSocketDenialResponse(
-                status_code=status_code,
-                headers=headers,
-                content=b"".join(body),
-            )
-
-    def send(self, message: Message) -> None:
-        self._receive_queue.put(message)
-
-    def send_text(self, data: str) -> None:
-        self.send({"type": "websocket.receive", "text": data})
-
-    def send_bytes(self, data: bytes) -> None:
-        self.send({"type": "websocket.receive", "bytes": data})
-
-    def send_json(self, data: typing.Any, mode: typing.Literal["text", "binary"] = "text") -> None:
-        text = json.dumps(data, separators=(",", ":"), ensure_ascii=False)
-        if mode == "text":
-            self.send({"type": "websocket.receive", "text": text})
-        else:
-            self.send({"type": "websocket.receive", "bytes": text.encode("utf-8")})
-
-    def close(self, code: int = 1000, reason: str | None = None) -> None:
-        self.send({"type": "websocket.disconnect", "code": code, "reason": reason})
-
-    def receive(self) -> Message:
-        message = self._send_queue.get()
-        if isinstance(message, BaseException):
-            raise message
-        return message
-
-    def receive_text(self) -> str:
-        message = self.receive()
-        self._raise_on_close(message)
-        return typing.cast(str, message["text"])
-
-    def receive_bytes(self) -> bytes:
-        message = self.receive()
-        self._raise_on_close(message)
-        return typing.cast(bytes, message["bytes"])
-
-    def receive_json(self, mode: typing.Literal["text", "binary"] = "text") -> typing.Any:
-        message = self.receive()
-        self._raise_on_close(message)
-        if mode == "text":
-            text = message["text"]
-        else:
-            text = message["bytes"].decode("utf-8")
-        return json.loads(text)
 
 
 class _TestClientTransport(httpx.BaseTransport):
@@ -336,7 +173,7 @@ class _TestClientTransport(httpx.BaseTransport):
                 "extensions": {"websocket.http.response": {}},
             }
             session = WebSocketTestSession(self.app, scope, self.portal_factory)
-            raise _Upgrade(session)
+            raise UpgradeException(session)
 
         scope = {
             "type": "http",
@@ -794,7 +631,7 @@ class TestClient(httpx.Client):
         kwargs["headers"] = headers
         try:
             super().request("GET", url, **kwargs)
-        except _Upgrade as exc:
+        except UpgradeException as exc:
             session = exc.session
         else:
             raise RuntimeError("Expected WebSocket upgrade")  # pragma: no cover
