+++ lilya/testclient/exceptions.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from lilya.testclient.base import WebSocketTestSession
+    from lilya.testclient.websockets import WebSocketTestSession
 
 
 class UpgradeException(Exception):
